+++ src/extensions/react-flow/nodes/NodeContent.tsx
@@ -4,15 +4,16 @@ import { useStore as getStoreStateFlowV12 } from "@xyflow/react";
 import Color from "color";
 import { NumberSize, Resizable, ResizableProps, ResizeCallback } from "re-resizable";
 
+import { preventReactFlowDragClass } from "../../../cmem/react-flow/ReactFlow/constants";
 import { intentClassName, IntentTypes } from "../../../common/Intent";
 import { Depiction, DepictionProps } from "../../../components";
 import { ValidIconName } from "../../../components/Icon/canonicalIconNames";
 import Icon from "../../../components/Icon/Icon";
 import OverflowText from "../../../components/Typography/OverflowText";
 import { CLASSPREFIX as eccgui } from "../../../configuration/constants";
+import { HandleDefault, HandleDefaultProps } from "../handles/HandleDefault";
 import { ReacFlowVersionSupportProps, ReactFlowVersions, useReactFlowVersion } from "../versionsupport";
 
-import { HandleDefault, HandleDefaultProps } from "../handles/HandleDefault";
 import { NodeContentExtensionProps } from "./NodeContentExtension";
 import { NodeDefaultProps } from "./NodeDefault";
 import { NodeHighlightColor } from "./sharedTypes";
@@ -649,7 +650,7 @@ export function NodeContent<CONTENT_PROPS = React.HTMLAttributes<HTMLElement>>({
                         )}
                     </div>
                     {(menuButtons || (showExecutionButtons && executionButtons)) && (
-                        <div className={`${eccgui}-graphviz__node__header-menu nodrag`}>
+                        <div className={`${eccgui}-graphviz__node__header-menu ${preventReactFlowDragClass}`}>
                             {showExecutionButtons && typeof executionButtons === "function"
                                 ? executionButtons(adjustedContentProps, setAdjustedContentProps)
                                 : null}
@@ -782,7 +783,7 @@ export function NodeContent<CONTENT_PROPS = React.HTMLAttributes<HTMLElement>>({
                     (resizeDirections!.right ? ` ${eccgui}-graphviz__node__resizer--right` : "") +
                     (resizeDirections!.bottom ? ` ${eccgui}-graphviz__node__resizer--bottom` : "")
                 }
-                handleWrapperClass={`${eccgui}-graphviz__node__resizer--cursorhandles` + " nodrag"}
+                handleWrapperClass={`${eccgui}-graphviz__node__resizer--cursorhandles ${preventReactFlowDragClass}`}
                 size={resizableSize}
                 maxHeight={resizeMaxDimensions?.height ?? undefined}
                 maxWidth={resizeMaxDimensions?.width ?? undefined}