+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+### Added
+
+- `<Switch />`
+    - `noDrag` parameter: Add the `nodrag` class to the Switch element. Default: `true`
+
+### Deprecated
+
+- `preventReactFlowActionsClasses`: use `ClassNames.ReactFlow.preventAllActions`
+
 ## [26.1.0] - 2026-08-20
 
 ### Added